requirements.txt
bottle==0.12.13
bottlenose=1.1.2
line-bot-sdk==1.3.0
beautifulsoup4==4.6.0
